Export json beatmap should be the the debug part.
